+++ lib/src/paperplane/mode/webhook.dart
@@ -3,8 +3,12 @@
  * Copyright (c) 2020 Nebulino
  */
 
+import 'package:paperplane/paperplane_exceptions.dart';
+
 /// It helps creating a Webhook bot.
 /// TODO: implement webhook
 class Webhook {
-  Webhook();
+  Webhook() {
+    throw PaperPlaneException(description: 'Not yet implemented.');
+  }
 }