+++ lib/src/telegram/core.dart
@@ -12,7 +12,9 @@ import 'package:paperplane/telegram.dart';
 
 part 'core.g.dart';
 
+part '../telegram/core/bot_command.dart';
 part '../telegram/core/callback_query.dart';
+part '../telegram/core/dice.dart';
 part '../telegram/core/login_url.dart';
 part '../telegram/core/response_parameters.dart';
 part '../telegram/core/update.dart';